Changed inconsistencies on spanish translation
There was a part that said 'llavero' for the literal translation to spanish of 'keyring'. Down more lines it said 'keyring' and not the the literal translation, as a native spanish speaker i think it would be easier by the context to determine what is 'keyring', 'llavero' would be sort of confusing i think. On a another line, there was some missing translation for 'ranking mirrors'. I changed it to 'clasificando espejos' following the consistency of the other translation. I went through the file a few times and checked that 'llavero' was changed to 'keyring' and 'mirrors' to 'espejos'.